fix: resolve build errors - decimal/double conversion and missing using directive
- Add LavaMeuCarro.Domain.Entities using to NotificationsController
- Fix decimal to double casting in PromotionsController Haversine calculation
- Change PromotionDTO latitude/longitude from double? to decimal? to match Unidade entity

+++ LavaMeuCarroApi/src/LavaMeuCarro.API/Controllers/PromotionsController.cs
@@ -34,7 +34,11 @@ public class PromotionsController : ControllerBase
             // Filter by location if coordinates provided
             if (lat.HasValue && lng.HasValue && radius.HasValue && unidade != null && unidade.Latitude.HasValue && unidade.Longitude.HasValue)
             {
-                var distance = CalculateDistance(lat.Value, lng.Value, unidade.Latitude.Value, unidade.Longitude.Value);
+                var distance = CalculateDistance(
+                    lat.Value, 
+                    lng.Value, 
+                    (double)unidade.Latitude.Value, 
+                    (double)unidade.Longitude.Value);
                 if (distance > radius.Value)
                     continue; // Skip this promotion - out of radius
             }
@@ -93,6 +97,6 @@ public record PromotionDTO(
     string UnidadeCity,
     decimal? AverageRating,
     string? UnidadeLogoUrl,
-    double? UnidadeLatitude = null,
-    double? UnidadeLongitude = null
+    decimal? UnidadeLatitude = null,
+    decimal? UnidadeLongitude = null
 );
